chore: move Portal component to root components folder
