+++ src/lib/components/interactive/UserButton/UserButtonUserProfilePage.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
 	import { getContext, onMount, mount, type Snippet, unmount, type Component } from 'svelte';
 	import type { UserButtonContext } from './types';
-	import Portal from './Portal.svelte';
+	import Portal from '$lib/components/Portal.svelte';
 
 	const { addCustomPage } = getContext<UserButtonContext>('$$_userButton');
